move py stuff up
