+++ src/common/error/Cargo.toml
@@ -8,9 +8,19 @@ edition = { workspace = true }
 serde_json = "1.0.96"
 
 
+[dependencies.pyo3]
+features = ["extension-module", "abi3-py37"]
+optional = true
+version = "0.18.3"
+
 [dependencies.arrow2]
 branch = "clark/expand-casting-support"
 features = ["chrono-tz", "compute_take", "compute_cast", "compute_aggregate", "compute_if_then_else", "compute_sort", "compute_filter", "compute_temporal", "compute_comparison", "compute_arithmetics", "compute_concatenate", "io_ipc"]
 git = "https://github.com/Eventual-Inc/arrow2"
 package = "arrow2"
 version = "0.17.1"
+
+
+[features]
+default = ["python"]
+python = ["dep:pyo3"]
