+++ src/core/src/python/error.rs
@@ -5,14 +5,3 @@ use pyo3::{
 
 use common_error::DaftError;
 
-impl std::convert::From<DaftError> for PyErr {
-    fn from(err: DaftError) -> PyErr {
-        match err {
-            DaftError::PyO3Error(pyerr) => pyerr,
-            DaftError::FileNotFound { path, source } => {
-                PyFileNotFoundError::new_err(format!("File: {path} not found\n{source}"))
-            }
-            _ => PyValueError::new_err(err.to_string()),
-        }
-    }
-}
